+++ src/borg/helpers/parseformat.py
@@ -1330,9 +1330,12 @@ class BorgJsonEncoder(json.JSONEncoder):
 def basic_json_data(manifest, *, cache=None, extra=None):
     key = manifest.key
     data = extra or {}
-    data |= {"repository": BorgJsonEncoder().default(manifest.repository), "encryption": {"mode": key.ARG_NAME}}
+    data |= {
+        "repository": BorgJsonEncoder().default(manifest.repository),
+        "encryption": {"encryption": key.ENC_NAME, "id_hash": key.IDHASH_NAME},
+    }
     data["repository"]["last_modified"] = OutputTimestamp(manifest.last_timestamp)
-    if key.NAME.startswith("key file"):
+    if getattr(key, "storage", None) == KeyBlobStorage.KEYFILE:
         data["encryption"]["keyfile"] = key.find_key()
     if cache:
         data["cache"] = cache